requirements.txt
pymysql
python-dotenv
fpdf2
pyodbc
